refactor: 클래스 이름 및 경로 변경 #36
